refactor: move DockArea::show_separator to separate file
